requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
websockets>=11.0.3
PyJWT==2.8.0
aiofiles==23.2.1
httpx==0.25.2
asyncpg==0.29.0